feat : validation and users api

+++ routes/users.js
@@ -6,10 +6,14 @@ const bcrypt = require('bcrypt');
 const { stringify } = require('jade/lib/utils');
 
 // Get All Users
-router.get('/get-all', async function (req, res) {
+router.get('/get-all', async function (_req, res) {
   const users = await prisma.user.findMany();
+  if (users.length === 0 || users === null || users === undefined) {
+    res.send('Data Users is Empty');
+  }
   res.send(users);
 });
+
 
 // Get User by ID
 router.get('/get-user/:id', async function (req, res) {
@@ -19,22 +23,30 @@ router.get('/get-user/:id', async function (req, res) {
       id: parseInt(id),
     },
   });
+  if (user.length === 0 || user === null || user === undefined) {
+    res.status(404).json('No Users Found');
+  }else{
+  res.json('user with id ${id} not found');
+  }
   res.send(user);
 });
+
 
 // Create User
 router.post('/create', async function (req, res) {
   const { name, email, password } = req.body;
-  const hashPassword = await bcrypt.hash(password, 10);
-  const stringPassword = await stringify(hashPassword);
-  const user = await prisma.user.create({
-    data: {
-      username: name,
-      email,
-      password: stringPassword,
-    },
+  name === '' ? res.json('Please fill the name field') : email === '' ? res.json('Please fill the email field') : password === '' ? res.json('Please fill the password field') : (async () => {
+    const hashPassword = await bcrypt.hash(password, 10);
+    const stringPassword = stringify(hashPassword);
+    const user = await prisma.user.create({
+      data: {
+        username: name,
+        email,
+        password: stringPassword,
+      },
+    });
+    res.send(user);
   });
-  res.send(user);
 });
 
 // Update User
@@ -43,28 +55,41 @@ router.put('/update/:id', async function (req, res) {
   const { name, email, password } = req.body;
   const hashPassword = bcrypt.hash(password, 10);
   const stringPassword = stringify(hashPassword);
-  const user = await prisma.user.update({
-    where: {
-      id: parseInt(id),
-    },
-    data: {
-      username: name,
-      email,
-      password: stringPassword,
-    },
-  });
-  res.send(user);
+  name === '' ? res.json('Please fill the name field') : email === '' ? res.json('Please fill the email field') : password === '' ? res.json('Please fill the password field') : (async () => {
+    const user = await prisma.user.update({
+      where: {
+        id: parseInt(id),
+      },
+      data: {
+        username: name,
+        email,
+        password: stringPassword,
+      },
+    });
+    res.send(user);
+  })
 });
+
 
 // Delete User
 router.delete('/delete/:id', async function (req, res) {
   const { id } = req.params;
-  const user = await prisma.user.delete({
+  const userExist = await prisma.user.findUnique({
     where: {
-      id: parseInt(id),
-    },
-  });
-  res.send(user);
+      id: parseInt(id)
+    }
+  })
+
+  if(!userExist) {
+    res.status(404).json("User not found")
+  } else {
+    const user = await prisma.user.delete({
+      where: {
+        id: parseInt(id),
+      },
+    });
+    res.send(users);
+  }
 });
 
 module.exports = router;
